Change PmFetch to create structs at time of fetch
I prefer this for the API. Data is just data on the struct instead of
a method call. It makes it easier to reason about the lifecycle of the
structs too.

This also includes a change to copy the pmValueBlock pointer from the
pmValue. Using circular references to avoid the C pmResult being free'ed
doesn't work as expected. The finalizer if not granted to run for these
 circular references so this would cause a memory leak.