Change pmInDom and pmID to uint32
It turns out (at least on my Go install) that int is not the same size as
what GCC uses for size of int. The assumption that they both were
the size of the architecture is wrong. In Go its 64 bits but for my GCC,
its 32 bits on a 64 bit machine. This causes weird downstream issues
when indexing into a pointer of a slice passed into a C function.

I'm doing this for now, but a better solution might be to populate C
vectors in C code instead of passing in the data section of slices.